code/chapter07/official_demo/pyproject.toml
[project]
name = "official-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Alwaysungry"}
]
requires-python = ">=3.13, <3.15"
dependencies = [
    "hello-agents (>=0.2.8,<0.3.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "huggingface-hub (>=1.3.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
